refactor: extract helper function creating link from text
